# specs/001-secure-file-upload/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CAMS File Upload Service
  description: |-
    API for initiating and tracking file uploads in the CAMS processing pipeline.
  version: 1.2.0
servers:
  - url: https://api.cams.com/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development Server
tags:
  - name: FileUpload
    description: Operations for file upload initiation and status tracking.
  - name: Validation
    description: Header validation template management and error reporting.

paths:
  /uploads:
    post:
      tags:
        - FileUpload
      summary: Request a pre-signed URL for file upload
      description: |-
        Initiates the file upload process. The client provides metadata about the file,
        and the service returns a unique `fileId` and a secure, time-limited (15-minute)
        pre-signed URL to which the file can be uploaded directly to the quarantine bucket.
        After uploading, the client MUST call POST /uploads/{fileId}/confirm.
      operationId: requestUploadUrl
      requestBody:
        description: Metadata for the file to be uploaded.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
      responses:
        '200':
          description: Successfully generated upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad Request. The request is missing required fields or is malformed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /uploads/{fileId}/confirm:
    post:
      tags:
        - FileUpload
      summary: Confirm that the file upload is complete
      description: |-
        Called by the client after successfully uploading the file bytes to the pre-signed URL.
        This transitions the file record from AWAITING_UPLOAD to UPLOADED and publishes a
        FileReceivedEvent to trigger the malware scanning pipeline (Epic 2).
      operationId: confirmUpload
      parameters:
        - name: fileId
          in: path
          required: true
          description: The unique file identifier returned by POST /uploads
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Upload confirmed. Scanning pipeline triggered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmUploadResponse'
        '404':
          description: No file record found for the given fileId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /uploads/{fileId}/validation-errors:
    get:
      tags:
        - Validation
      summary: Get validation errors for a file
      description: |-
        Returns the header validation result and all field-level errors for the given file.
        Returns 404 if no validation run has been completed for this file yet.
        Returns 200 with an empty errors array if validation passed (PASS status).
      operationId: getValidationErrors
      parameters:
        - name: fileId
          in: path
          required: true
          description: The unique file identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Validation result found (may contain empty errors list for PASS).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: No validation result found for this fileId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /validation-templates:
    post:
      tags:
        - Validation
      summary: Create a new validation template
      description: |-
        Creates a new validation template for the given flowType.
        Any existing active template for the same flowType is deactivated.
        New template takes effect immediately (no restart required).
      operationId: createValidationTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationTemplateRequest'
      responses:
        '201':
          description: Template created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationTemplateResponse'
        '400':
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /validation-templates/{flowType}:
    put:
      tags:
        - Validation
      summary: Update column rules for a flow type
      description: |-
        Replaces the column rules for the given flowType. Increments version.
        Previous version is deactivated. New rules take effect immediately.
      operationId: updateValidationTemplate
      parameters:
        - name: flowType
          in: path
          required: true
          schema:
            type: string
          example: NAV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationTemplateRequest'
      responses:
        '200':
          description: Template updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationTemplateResponse'
        '400':
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict — template referenced by in-flight files cannot be deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Validation
      summary: Get the active validation template for a flow type
      operationId: getValidationTemplate
      parameters:
        - name: flowType
          in: path
          required: true
          schema:
            type: string
          example: NAV
      responses:
        '200':
          description: Active template found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationTemplateResponse'
        '404':
          description: No active template exists for this flowType.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    UploadRequest:
      type: object
      description: Body for requesting a file upload URL.
      required:
        - fileName
        - flowType
        - checksum
      properties:
        fileName:
          type: string
          description: The name of the file to be uploaded.
          example: "transactions_20260315.csv"
        flowType:
          type: string
          description: The business process category for which the file is being uploaded.
          example: "NAV"
        checksum:
          type: string
          description: |
            The MD5 (32 hex chars) or SHA-256 (64 hex chars) checksum of the file
            for integrity verification post-upload.
          pattern: '^[a-fA-F0-9]{32}$|^[a-fA-F0-9]{64}$'
          example: "d41d8cd98f00b204e9800998ecf8427e"

    UploadResponse:
      type: object
      description: Response containing the fileId and the pre-signed URL for upload.
      properties:
        fileId:
          type: string
          format: uuid
          description: A unique identifier for this file upload process.
          example: "123e4567-e89b-12d3-a456-426614174000"
        uploadUrl:
          type: string
          format: uri
          description: |
            The secure, 15-minute time-limited URL to which the client should upload
            the file via HTTP PUT. Call POST /uploads/{fileId}/confirm after upload.
          example: "https://storage.googleapis.com/cams-quarantine-bucket/..."

    ConfirmUploadResponse:
      type: object
      description: Response confirming that the upload has been acknowledged and scanning triggered.
      properties:
        fileId:
          type: string
          format: uuid
          description: The unique identifier of the confirmed file.
        status:
          type: string
          description: The new status of the file record (always UPLOADED on success).
          example: "UPLOADED"
        confirmedAt:
          type: string
          format: date-time
          description: Timestamp of the confirmation.

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # -------------------------------------------------------
    # Epic 3 — Header Validation Schemas
    # -------------------------------------------------------

    ValidationErrorResponse:
      type: object
      description: Header validation result including all field-level errors.
      properties:
        fileId:
          type: string
          format: uuid
        flowType:
          type: string
          example: NAV
        templateVersion:
          type: integer
        status:
          type: string
          enum: [PASS, FAIL, UNPARSEABLE_HEADER]
        errorCount:
          type: integer
        validatedAt:
          type: string
          format: date-time
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'

    ValidationErrorDetail:
      type: object
      properties:
        columnName:
          type: string
          nullable: true
        position:
          type: integer
          description: Zero-indexed column position (-1 if unknown)
        errorCode:
          type: string
          enum:
            - MISSING_COLUMN
            - WRONG_POSITION
            - EXTRA_COLUMN
            - UNPARSEABLE_HEADER
            - NULL_NOT_ALLOWED
            - INVALID_DATA_TYPE
        detail:
          type: string

    ValidationTemplateRequest:
      type: object
      required: [flowType, columnRules]
      properties:
        flowType:
          type: string
          example: NAV
        columnRules:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ColumnRuleRequest'

    ColumnRuleRequest:
      type: object
      required: [name, position, dataType]
      properties:
        name:
          type: string
          example: FUND_CODE
        position:
          type: integer
          minimum: 0
        dataType:
          type: string
          enum: [STRING, INTEGER, DECIMAL, DATE, UUID]
        required:
          type: boolean
          default: true
        allowNull:
          type: boolean
          default: false
        maxLength:
          type: integer
          nullable: true
        pattern:
          type: string
          nullable: true

    ValidationTemplateResponse:
      type: object
      properties:
        id:
          type: string
        flowType:
          type: string
        version:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        columnRules:
          type: array
          items:
            $ref: '#/components/schemas/ColumnRuleRequest'